use dateparser to parse the date to datetime
