--- etk/extractors/date_etl/timezone_abbr.txt
@@ -1,198 +1,0 @@
-ACDT
-ACST
-ACT
-ACT
-ACWST
-ADT
-AEDT
-AEST
-AFT
-AKDT
-AKST
-AMST
-AMT
-AMT
-ART
-AST
-AST
-AWST
-AZOST
-AZOT
-AZT
-BDT
-BIOT
-BIT
-BOT
-BRST
-BRT
-BST
-BST
-BST
-BTT
-CAT
-CCT
-CDT
-CDT
-CEST
-CET
-CHADT
-CHAST
-CHOT
-CHOST
-CHST
-CHUT
-CIST
-CIT
-CKT
-CLST
-CLT
-COST
-COT
-CST
-CST
-CST
-CT
-CVT
-CWST
-CXT
-DAVT
-DDUT
-DFT
-EASST
-EAST
-EAT
-ECT
-ECT
-EDT
-EEST
-EET
-EGST
-EGT
-EIT
-EST
-FET
-FJT
-FKST
-FKT
-FNT
-GALT
-GAMT
-GET
-GFT
-GILT
-GIT
-GMT
-GST
-GST
-GYT
-HDT
-HAEC
-HST
-HKT
-HMT
-HOVST
-HOVT
-ICT
-IDT
-IOT
-IRDT
-IRKT
-IRST
-IST
-IST
-IST
-JST
-KGT
-KOST
-KRAT
-KST
-LHST
-LHST
-LINT
-MAGT
-MART
-MAWT
-MDT
-MET
-MEST
-MHT
-MIST
-MIT
-MMT
-MSK
-MST
-MST
-MUT
-MVT
-MYT
-NCT
-NDT
-NFT
-NPT
-NST
-NT
-NUT
-NZDT
-NZST
-OMST
-ORAT
-PDT
-PET
-PETT
-PGT
-PHOT
-PHT
-PKT
-PMDT
-PMST
-PONT
-PST
-PST
-PYST
-PYT
-RET
-ROTT
-SAKT
-SAMT
-SAST
-SBT
-SCT
-SDT
-SGT
-SLST
-SRET
-SRT
-SST
-SST
-SYOT
-TAHT
-THA
-TFT
-TJT
-TKT
-TLT
-TMT
-TRT
-TOT
-TVT
-ULAST
-ULAT
-USZ1
-UTC
-UYST
-UYT
-UZT
-VET
-VLAT
-VOLT
-VOST
-VUT
-WAKT
-WAST
-WAT
-WEST
-WET
-WIT
-WST
-YAKT
-YEKT
